ingestion/spark_ingestion_job/requirements.txt
pyspark==3.5.0
requests
pyyaml
delta-spark
tenacity
logging
typing
pathlib